Complete separating interface/implementation in all mapping/* files.
