+++ src/SConscript
@@ -109,6 +109,8 @@ loop-analysis/operation-type.cpp
 pat/pat.cpp
 mapping/loop.cpp
 mapping/nest.cpp
+mapping/arch-properties.cpp
+mapping/constraints.cpp
 model/arithmetic.cpp
 model/buffer.cpp
 model/topology.cpp
